+++ internal/protocol/generated.go
@@ -771,7 +771,7 @@ type InitialStateMessage struct {
 	Sessions []Session `json:"sessions,omitempty"`
 
 	// Settings corresponds to the JSON schema field "settings".
-	Settings map[string]interface{} `json:"settings,omitempty"`
+	Settings RecordString `json:"settings,omitempty"`
 
 	// Warnings corresponds to the JSON schema field "warnings".
 	Warnings []DaemonWarning `json:"warnings,omitempty"`
@@ -1156,6 +1156,8 @@ type RecentLocationsResultMessage struct {
 	Success bool `json:"success"`
 }
 
+type RecordString map[string]interface{}
+
 type RefreshPRsMessage struct {
 	// Cmd corresponds to the JSON schema field "cmd".
 	Cmd string `json:"cmd"`
@@ -1419,7 +1421,7 @@ type ReviewState struct {
 
 type ReviewToolUse struct {
 	// Input corresponds to the JSON schema field "input".
-	Input map[string]interface{} `json:"input"`
+	Input RecordString `json:"input"`
 
 	// Name corresponds to the JSON schema field "name".
 	Name string `json:"name"`
@@ -1470,6 +1472,9 @@ type Session struct {
 	// NeedsReviewAfterLongRun corresponds to the JSON schema field
 	// "needs_review_after_long_run".
 	NeedsReviewAfterLongRun *bool `json:"needs_review_after_long_run,omitempty"`
+
+	// Recoverable corresponds to the JSON schema field "recoverable".
+	Recoverable *bool `json:"recoverable,omitempty"`
 
 	// State corresponds to the JSON schema field "state".
 	State SessionState `json:"state"`
@@ -1580,7 +1585,7 @@ type SettingsUpdatedMessage struct {
 	Event string `json:"event"`
 
 	// Settings corresponds to the JSON schema field "settings".
-	Settings map[string]interface{} `json:"settings,omitempty"`
+	Settings RecordString `json:"settings,omitempty"`
 
 	// Success corresponds to the JSON schema field "success".
 	Success *bool `json:"success,omitempty"`
@@ -1955,7 +1960,7 @@ type WebSocketEvent struct {
 	Sessions []Session `json:"sessions,omitempty"`
 
 	// Settings corresponds to the JSON schema field "settings".
-	Settings map[string]interface{} `json:"settings,omitempty"`
+	Settings RecordString `json:"settings,omitempty"`
 
 	// Signal corresponds to the JSON schema field "signal".
 	Signal *string `json:"signal,omitempty"`